parser/htmlparser/tests/mochitest/test_bug655682.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=655682
-->
<head>
  <title>Test for Bug 655682</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=655682">Mozilla Bug 655682</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe src=file_bug655682.sjs></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 655682 **/

var tdsSeen = 0;

var triggeredSecondTd = false;

var iframe = document.getElementsByTagName("iframe")[0];

SimpleTest.waitForExplicitFinish();

function probe() {
  var tds = iframe.contentDocument.getElementsByTagName("td").length;
  switch (tds) {
    case 0:
      setTimeout(probe, 0);
      return;
    case 1:
      tdsSeen = tds;
      if (!triggeredSecondTd) {
        triggeredSecondTd = true;
        var script = document.createElement("script");
        script.src = "file_bug655682.sjs?trigger=1";
        document.head.appendChild(script);
      }
      setTimeout(probe, 0);
      return;
    case 2:
      is(tdsSeen, 1, "Should have seen one td before seeing two.");      
      SimpleTest.finish();
      return;
    default:
      ok(false, "Wrong number of tds");
      SimpleTest.finish();
  }
}

setTimeout(probe, 0);

</script>
</pre>
</body>
</html>
